Rename and merge credit note writer
